update test script

+++ tests/test_pytorch_transformers_interface.py
@@ -40,7 +40,7 @@ class TestPytorchTransformersInterface(unittest.TestCase):
         # [CLS] 8 9 [SEP] 10 [SEP] [PAD]
         inp = torch.Tensor([[5, 8, 3, 9, 10, 11, 3], [5, 8, 9, 3, 10, 3, 7]])
         mask = inp != bert_model._pad_id
-        output = bert_model.get_seg_ids(bert_model, inp, mask)
+        output = bert_model.get_seg_ids(bert_model, inp, mask.long())
         assert torch.all(
             torch.eq(output, torch.Tensor([[0, 0, 0, 1, 1, 1, 1], [0, 0, 0, 0, 1, 1, 0]]))
         )
@@ -49,7 +49,7 @@ class TestPytorchTransformersInterface(unittest.TestCase):
         # [CLS] 8 [SEP] [PAD]
         inp = torch.Tensor([[5, 8, 9, 3], [5, 9, 3, 7]])
         mask = inp != bert_model._pad_id
-        output = bert_model.get_seg_ids(bert_model, inp, mask)
+        output = bert_model.get_seg_ids(bert_model, inp, mask.long())
         assert torch.all(torch.eq(output, torch.Tensor([[0, 0, 0, 0], [0, 0, 0, 0]])))
 
     def test_xlnet_seg_ids(self):
@@ -65,7 +65,7 @@ class TestPytorchTransformersInterface(unittest.TestCase):
         # 8 [SEP] 9 10 [SEP] [CLS] [PAD]
         inp = torch.Tensor([[8, 3, 9, 10, 11, 3, 5], [8, 3, 9, 10, 3, 5, 7]])
         mask = inp != xlnet_model._pad_id
-        output = xlnet_model.get_seg_ids(xlnet_model, inp, mask)
+        output = xlnet_model.get_seg_ids(xlnet_model, inp, mask.long())
         assert torch.all(
             torch.eq(output, torch.Tensor([[0, 3, 1, 1, 1, 3, 2], [0, 3, 1, 1, 3, 2, 0]]))
         )
@@ -74,5 +74,5 @@ class TestPytorchTransformersInterface(unittest.TestCase):
         # 8 9 [SEP] [CLS] [PAD]
         inp = torch.Tensor([[8, 9, 10, 3, 5], [8, 9, 3, 5, 7]])
         mask = inp != xlnet_model._pad_id
-        output = xlnet_model.get_seg_ids(xlnet_model, inp, mask)
+        output = xlnet_model.get_seg_ids(xlnet_model, inp, mask.long())
         assert torch.all(torch.eq(output, torch.Tensor([[0, 0, 0, 3, 2], [0, 0, 3, 2, 0]])))
